Move harvest create-only stash writes into stash.
Keeps diverge skip/repair intent with the stash owner so harvest no longer reimplements preflight branches.
